Fixes a memory leak in the Escape keyup handler
Previously, the super-modal component's "this.listener" was set to the return value of "document.addEventListener()" rather than the actual keyup handler. This meant that on "willDestroyElement" the listener was not actually removed. This caused a) a memory leak, and b) errors in the consuming app, when consuming component's "onClose" handlers were called after the components had already been destroyed.